Updated ALL SONGS. (desc)
I updated ALL of the songs in the mod which basically is just getting all songs to have capital letters and fix the difficulties/add difficulties that are supposed to be in the mod + Giving song IDs so we know which song is which. Also deleting scripts that don't really need to be in here. But This freed up ALOT of memory so FPS is very crispy now I also made the experience better.

It may look like the songs have gotten smaller, but it's the same songs and like 2 songs got replaced.