Last readme update till separate engine repo
Psych Ward told me it wasn't a psych fork only cause of "stolen code" when it was simply clear that some of the code can be used!

Anyway the code wasn't really used much so these are the current stuff before I separate the mod with its own base engine. 

Please and please be aware I'm still learning. I did forget somethings which is fine.
I'll instead make it its own engine than a Psych fork (which was the original plan anyway).