Finished reverseScroll (still somewhat buggy) with OD and HP soon
It works perfectly! There will be some crashes as it isn't perfect. But i hope this doesn't cause any problems in the future. We got a cool coder helping out behind the scenes plus some coders from Mixtape for some changes.

It is the summer, sdypj is back in progress!